requirements.txt
cloudpickle==0.5.3
cycler==0.10.0
Cython==0.28.3
dask==0.17.5
decorator==4.3.0
h5py==2.8.0
imageio==2.3.0
kiwisolver==1.0.1
matplotlib==2.2.2
networkx==2.1
numpy==1.14.4
Pillow==5.1.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
scikit-image==0.14.0
scipy==1.1.0
six==1.11.0
toolz==0.9.0
torch
torchvision
